pyproject.toml
[project]
name = "jedisos"
version = "0.1.1"
description = "AI Agent System with zvecsearch Memory + LangGraph + LiteLLM"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "JediSOS Team" }]

dependencies = [
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    "litellm>=1.81.12",
    "langgraph>=1.0.8",
    "langgraph-checkpoint-postgres>=3.0.4",
    "argon2-cffi>=23.1.0",
    "fastmcp>=2.14.5,<3.0",
    "mcp>=1.26.0,<2.0",
    "httpx>=0.28.1",
    "structlog>=25.5.0",
    "typer[all]>=0.23.1",
    "rich>=14.3.2",
    "uuid6>=2025.0.1",
    "croniter>=6.0.0",
    "cryptography>=46.0.5",
    "keyring>=25.7.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.2",
    "ddgs>=8.0.0",
]

[project.optional-dependencies]
dspy = [
    "dspy>=3.1.0",
]
channels = [
    "python-telegram-bot>=22.6",
    "discord.py>=2.6.4",
    "slack-bolt>=1.27.0",
]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-timeout>=2.4.0",
    "pytest-xdist>=3.5",
    "hypothesis>=6.151.8",
    "ruff>=0.15.1",
    "bandit[toml]>=1.9.3",
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.1",
]

[project.scripts]
jedisos = "jedisos.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jedisos"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["jedisos"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: 통합 테스트 (실제 Hindsight 필요)",
    "e2e: E2E 테스트 (전체 스택 필요)",
]
timeout = 30

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]

[tool.coverage.run]
source = ["jedisos"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 70
show_missing = true
